client/requirements.txt
requests
python-keystoneclient
